requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
nltk
spacy
requests
flask
gunicorn
openai